requirements.txt
langchain
streamlit
faiss-cpu
sentence_transformers
requests
